requirements.txt
# Core Dependencies
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.2
websocket-client>=1.6.0
ccxt>=4.0.0

# WebSocket & Async
websockets>=12.0
aiohttp>=3.9.0
asyncio>=3.4.3

# Scheduling & Background Tasks
schedule>=1.2.0
APScheduler>=3.10.0
python-crontab>=2.7.1
celery>=5.3.0

# Data Processing & Analysis
ta>=0.10.0  # Technical Analysis Library
scipy>=1.11.0
scikit-learn>=1.3.0
statsmodels>=0.14.0

# Database & Storage
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL
redis>=5.0.0
pymongo>=4.5.0  # MongoDB

# Utilities & Logging
python-dotenv>=1.0.0
colorlog>=6.7.0
tqdm>=4.65.0
pytest>=7.4.0
black>=23.0.0

# API & HTTP Clients
httpx>=0.25.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Monitoring & Alerting
prometheus-client>=0.19.0
slack-sdk>=3.21.0
twilio>=8.10.0

# Telegram Bot
python-telegram-bot>=20.7